Move `PacketLocation` into its own module
Being bundled with `OpenPGPMessage` is creating the need for horrid
circular imports with `decoder.py`
