+++ encryptit/openpgp_message.py
@@ -1,4 +1,5 @@
 from .compat import OrderedDict
+from .decoder import find_packets
 
 
 class OpenPGPMessage(object):
@@ -18,7 +19,6 @@ class OpenPGPMessage(object):
             print(f.serialize())  # f still in use
         ```
         """
-        from .decoder import find_packets  # unavoidable circular import
         return cls(packet_locations=find_packets(f))
 
     def __init__(self, packet_locations=None):
@@ -31,36 +31,3 @@ class OpenPGPMessage(object):
         return OrderedDict([
             ('packet_locations', self.packet_locations),
         ])
-
-
-class PacketLocation(object):
-
-    def __init__(self, header_start, body_start, body_length):
-        self.header_start = header_start
-        self.body_start = body_start
-        self.body_length = body_length
-
-    def serialize(self):
-        return OrderedDict([
-            ('header_start', self.header_start),
-            ('header_length', self.header_length),
-            ('header_end', self.header_end),
-            ('body_start', self.body_start),
-            ('body_length', self.body_length),
-            ('body_end', self.body_end),
-        ])
-
-    @property
-    def header_length(self):
-        return self.body_start - self.header_start
-
-    @property
-    def header_end(self):
-        return self.body_start
-
-    @property
-    def body_end(self):
-        return self.body_start + self.body_length
-
-    def __eq__(self, other):
-        return other.__dict__ == self.__dict__
